[TAJO-911] Refactoring MySQL/MariaDB Catalog Store
